Update quickstart.md
Added a check for troubleshooting "Nginx Unavailable" message. I had used the AWS VPC wizard to setup a VPC and security group, but the "docker-machine" security group did not have port 80 opened as part of that setup. This edit includes a check for that, in the event that an Nginx Unavailable message is observed.